+++ resources/leiningen/new/lein_compojure/queries.sql
@@ -1,0 +1,1 @@
+-- SQL for the {{humanized-name}} project. Refer to https://www.hugsql.org/ for usage.