+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. This change
 - Docs on how to use prone, run a focused test
 - Support for logfmt
 - `lein server` also start an nREPL that can be connected to with `lein repl :connect`
+- Optional PostgreSQL support w/ migrations
 
 ### Fixed
 - Use correct test assertion order in handler test